+++ app/src/Routes/AuthRoutes.php
@@ -17,6 +17,7 @@ use Slim\Routing\RouteCollectorProxy;
 use UserFrosting\Routes\RouteDefinitionInterface;
 use UserFrosting\Sprinkle\Account\Authenticate\AuthGuard;
 use UserFrosting\Sprinkle\Account\Authenticate\GuestGuard;
+use UserFrosting\Sprinkle\Account\Controller\AuthCheckAction;
 use UserFrosting\Sprinkle\Account\Controller\CaptchaAction;
 use UserFrosting\Sprinkle\Account\Controller\CheckUsernameAction;
 use UserFrosting\Sprinkle\Account\Controller\DenyResetPasswordAction;
@@ -56,6 +57,7 @@ class AuthRoutes implements RouteDefinitionInterface
 
         // No guard
         $app->group('/account', function (RouteCollectorProxy $group) {
+            $group->get('/auth-check', AuthCheckAction::class)->setName('account.authCheck');
             $group->get('/captcha', CaptchaAction::class)->setName('account.captcha');
             $group->get('/check-username', CheckUsernameAction::class)->setName('account.checkUsername');
             $group->get('/suggest-username', SuggestUsernameAction::class)->setName('account.suggestUsername');